+++ src/main/java/com/moa/moa_server/domain/notification/application/event/NotificationHandler.java
@@ -1,4 +1,4 @@
-package com.moa.moa_server.domain.notification.event;
+package com.moa.moa_server.domain.notification.application.event;
 
 import com.moa.moa_server.domain.notification.entity.Notification;
 import com.moa.moa_server.domain.notification.repository.NotificationRepository;
