+++ src/main/java/com/moa/moa_server/domain/notification/application/producer/GroupNotificationProducerImpl.java
@@ -1,10 +1,10 @@
-package com.moa.moa_server.domain.notification.producer;
+package com.moa.moa_server.domain.notification.application.producer;
 
 import com.moa.moa_server.domain.group.entity.Group;
 import com.moa.moa_server.domain.group.repository.GroupMemberRepository;
+import com.moa.moa_server.domain.notification.application.event.EventPublisher;
+import com.moa.moa_server.domain.notification.application.event.NotificationEvent;
 import com.moa.moa_server.domain.notification.entity.NotificationType;
-import com.moa.moa_server.domain.notification.event.EventPublisher;
-import com.moa.moa_server.domain.notification.event.NotificationEvent;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
