+++ src/main/java/com/moa/moa_server/domain/notification/application/producer/NotificationProducer.java
@@ -1,4 +1,4 @@
-package com.moa.moa_server.domain.notification.producer;
+package com.moa.moa_server.domain.notification.application.producer;
 
 /** 알림 생성자 역할을 명시하는 공통 인터페이스 */
 public interface NotificationProducer {}
